% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barMonophyletic.R
\name{barMonophyletic}
\alias{barMonophyletic}
\title{barMonophyletic}
\usage{
barMonophyletic(groupLabel, groupMatch, tree, cex.plot, cex.text = 0.8,
  include.tip.label = FALSE, extra.space = 0, coef.space = 1,
  draw = TRUE, text.offset = 1.02, font = 1, font.col = 1,
  seg.col = 1, srt = 0, bar.at.tips = FALSE)
}
\arguments{
\item{groupLabel}{a vector of mode \code{character} that indicates
the labels for each of the monophyletic group.}

\item{groupMatch}{a vector of mode \code{character} that contains
a way to identify each of the monophyletic clade. The function
\code{\link{grep}} is used for this process.}

\item{tree}{a phylogenetic tree a \code{phylo} object.}

\item{cex.plot}{the \code{\link[par]{cex}} (in \code{par}) value
for plotting the tree.}

\item{cex.text}{the \code{\link[par]{cex}} (in \code{par}) value
for the text next to the vertical bars.}

\item{include.tip.label}{should the tip labels be displayed on the
tree? (default is FALSE)}

\item{extra.space}{amount of extra space (additive)}

\item{coef.space}{amount of space (multiplicative)}

\item{draw}{should a plot be produce? (default is TRUE)}

\item{text.offset}{how much space to add between the vertical bar
and the text.}

\item{font}{type of font to be used for the text next to the
vertical bars.}

\item{font.col}{color of the font to be used for the text next to
the vertical bars.}

\item{seg.col}{color of the vertical bar.}

\item{srt}{the string rotation in degrees (see
\code{see \code{srt} in \link[graphics]{par}}) of the text next to the
vertical bars.}

\item{bar.at.tips}{should the vertical bars be plotted right next
to the tips? (default FALSE)}
}
\value{
Function mostly used for its side effect of plotting a phylogeny
but it also returns invisibly the total width of the tree.
}
\description{
puts vertical bars next to monophyletic clades in a tree.
}
\details{
Based on a tree (as a \code{phylo} object) and a character-vector,
this function draws vertical bars next to each monophyletic
group. Each element of the vector must contain a character string
that matches (through \code{\link{grep}}) all the tips that are
included in each of the monophyletic group. In other words, this
function does not identify the monophyletic groups for you. Also,
the length of the vector should be of the same length as the number
of monophyletic groups in your tree.

The element of the vector are considered labels and will be
displayed next to the vertical bars.
}
\author{
Francois Michonneau
}

