% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chopper-package.R
\docType{package}
\name{chopper}
\alias{chopper}
\alias{chopper-package}
\title{Sequence data manipulation to ease phylogenetic analyses}
\description{
A suite of tools to automate and facilitate data preparation for
phylogenetic and population genetic analyses.
}

