% Generated by roxygen2 (4.0.0): do not edit by hand
\name{completeSeqWithGaps}
\alias{completeSeqWithGaps}
\title{Complete sequences with gaps.}
\usage{
completeSeqWithGaps(file, output, format = "fasta", colsep = "",
  gaps = "-", ...)
}
\arguments{
\item{file}{input file in the FASTA format or other format that
can be read by ape:::read.dna}

\item{output}{path and name file for the output.}

\item{format}{format fo the input file, to be passed to
ape:::read.dna}

\item{colsep}{a character used to separate the columns}

\item{gaps}{character to be as gap}

\item{...}{further arguments to be passed to ape:::write.dna to
format the output file.}
}
\value{
TRUE. This function is intended to be use for its side
effect which create a new alignment file with standardized
sequence lengths.
}
\description{
Standardize the lengths of the sequences in a FASTA file by adding
gaps (or other characters) at the end.
}
\details{
This function takes an alignment in which the sequences are of
different lengths (for instance, if 2 alignments are concatenated
manually), and standardizes the lengths by adding gaps at the end
of the sequences. If you want the output file to be aligned, you
need to make sure that the sequences are otherwise aligned (they
only differ by some gaps at their extremities).
}
\author{
Francois Michonneau
}

