% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concatenateAlignments.R
\name{concatenateAlignments}
\alias{concatenateAlignments}
\title{Concatenate alignments}
\usage{
concatenateAlignments(pattern, path, output, input.format = "fasta",
  partition = NULL, partition.format = c("raxml", "nexus"),
  create.conc = TRUE, standardize = FALSE, drop = NULL, ...)
}
\arguments{
\item{pattern}{character string to identify aligned files in folder}

\item{path}{path where the alignment files are to be found}

\item{output}{file name (with full path) for the concatenated file}

\item{input.format}{format of the input DNA alignment (to be
passed to ape:::read.dna)}

\item{partition}{if NULL no partition file is created, otherwise a
character string indicating the file name where to write the
partitions}

\item{partition.format}{which format should be used to create the
partition file (raxml or nexus)}

\item{create.conc}{if TRUE create the concatenated file, if FALSE
useful for testing (dry run) or to create partition file}

\item{standardize}{not needed if alignments created by
\sQuote{mergeSeq}. Otherwise the alignments names are standardized
for each of the individual markers by generating empty sequences.}

\item{drop}{a character vector indicating the sequences that need
to be removed from the output}

\item{...}{further arguments to be passed to write.dna for the
output file}
}
\value{
This function is used for its side effect of creating an
alignment from a set of alignemnts (marker based). It can also
create a partition file that can added at the end of the alignment
to specify where each marker starts and ends.
}
\description{
Concenate marker alignements to generate a single alignment.
}
\details{
Given a series of files generated by \sQuote{mergeSeq}, this
function creates an alignment from these files. WARNING: the files
for the independent loci are assumed to be created using the
function mergeSeq, meaning that the loci is indicated in 3rd
position, separated by an hyphen e.g. 20120808-101010-COI.fas
}
\examples{
\dontrun{
concatenateAlignments(pattern="^20121017-112755.+afa$", path="/tmp/seq",
                      output="/tmp/seq/20121017.impatiens.fas",
                      partition="/tmp/seq/20121017.impatiens.part",
                      partition.format="nexus", colsep="", colw=10000)
}
}
\author{
Francois Michonneau
}

