% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertGaps.R
\name{convertGaps}
\alias{convertGaps}
\title{Convert gap characters}
\usage{
convertGaps(file, output, formatin = "sequential", formatout = formatin,
  colsep = "", from = "?", to = "-", overwrite = FALSE, ...)
}
\arguments{
\item{file}{path and file name of a DNA alignment}

\item{output}{path and file name for the modified file}

\item{formatin}{format of the input file (to be passed to
\code{\link[ape]{read.dna}}).}

\item{formatout}{format for the output file (to be passed to
\code{\link[ape]{write.dna}}, by default the same as
\code{formatin}).}

\item{colsep}{argument to be passed to
\code{\link[ape]{write.dna}}, character used to separate columns
in DNA alignment (\code{ape}'s default is a single space, here we
override this to no character, i.e. \code{colsep=""}).}

\item{from}{character to be changed in alignment (default \dQuote{?})}

\item{to}{new character to use in alignment (default \dQuote{-})}

\item{overwrite}{if no output file is specified, and
\code{overwrite} is TRUE, then the original file is overwritten.}

\item{...}{additional arguments to be passed to
\code{\link[ape]{write.dna}}.}
}
\value{
the name of the output file (invisibly).
}
\description{
Convert gap characters in an alignemnt
}
\details{
This function is intended to replace one type of gap character
(e.g., \sQuote{?}) into another (e.g., \sQuote{-}).
}
\author{
Francois Michonneau
}

