% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandPartitionInfo.R
\name{expandPartitionInfo}
\alias{expandPartitionInfo}
\title{expandPartitionInfo}
\usage{
expandPartitionInfo(partitionInfo)
}
\arguments{
\item{partitionInfo}{a partitionInfo object}
}
\value{
a list named with partition names, that contains all base
pair positions for each partition.
}
\description{
Expands information contained in a partitionInfo object to obtain
all base pair positions for a given alignment.
}
\details{
The partition are expanded in the same order as they are placed in
the partition file.
}
\examples{
cat("DNA, p1 = 1-15\\\\3,2-15\\\\3",
    "DNA, p2 = 3-15\\\\3", sep="\\n",
    file="testPartition.part")
pInfo <- raxmlPartitionInfo("testPartition.part")
expandPartitionInfo(pInfo)
unlink("testPartition.part")
}
\author{
Francois Michonneau
}
\seealso{
raxmlPartitionInfo
}

