% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeSequences.R
\name{mergeSeq}
\alias{mergeSeq}
\title{Merge sequences}
\usage{
mergeSeq(listFiles, output, seqFolder = "~/Documents/seqRepository",
  markers = c("16S", "16Sc"), convertEnds = TRUE, checkAmbiguity = TRUE,
  gblocks = NULL, gapchar = "?", justCheck = FALSE, returnData = TRUE)
}
\arguments{
\item{listFiles}{names (vector of character) of the individual
sequence files to be merged.}

\item{output}{folder where the write the output files. If the
folder doesn't exist, the user is asked whether it should be
created.}

\item{seqFolder}{the folder that contains the individual files
being stored in their own folders.}

\item{markers}{the list of markers to be used to build the
alignemnt. They need to match the subfolders found in
\sQuote{seqFolder}}

\item{convertEnds}{Should the gaps at the extremities of the
sequences be replaced by missing data?}

\item{checkAmbiguity}{Should the sequences be checked for
ambiguities before merging?}

\item{gblocks}{If different from NULL, gblocks should be a named
list with the marker as its names and the parameters to be passed
to Gblocks as its values. This function assumes that the extension
of the file created by Gblocks is \sQuote{-gb} so you don't want
to use the \sQuote{-e=} flag.}

\item{gapchar}{Characters to be used to create empty sequences for
files that are not sequenced for a given marker.}

\item{justCheck}{If FALSE, no alignment is created.}

\item{returnData}{If TRUE, a data frame that contains TRUE/FALSE
indicating whether a particular extraction (element of listFiles)
is found for each of the markers.}
}
\value{
Mostly used for its side effect of creating alignment for
a list of extractions, given a list of markers. If returnData is
TRUE, it also returns a data frame that includes which extraction
has been sequenced for which marker.
}
\description{
Creates a set of alignments for given a list of extractions and
markers.
}
\details{
This function takes a list of extractions and creates alignments
from them. When more than one marker is specified, empty sequences
might be added if they are not available for this particular
marker, to create alignments with the same number of sequences
across all markers. The alignment is being performed by
MUSCLE. This function can also call Gblocks to remove ambiguous
parts in the alignment. To function properly R needs to be able to
access MUSCLE and Gblocks with the command system(), this means
that the executables need to be in your PATH.
}
\author{
François Michonneau
}

