% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nbases.R
\name{nbase}
\alias{nbase}
\title{Number of bases}
\usage{
nbase(file, gap = "-", format = "fasta", ...)
}
\arguments{
\item{file}{A file with DNA sequences}

\item{gap}{The character(s) used as gaps in the file}

\item{format}{The file format of the alignment}

\item{...}{Additional parameters to be passed to \code{\link[ape]{read.dna}}}
}
\value{
a named vector indicating the number of non-gap characters found in each sequence.
}
\description{
Calculates the number of base pairs in each sequence of an alignment.
}
\details{
The argument \code{gap} should be a valid regular expression whose
value is substituted (by \code{gsub}) into empty characters. The
remaining is the base pairs making up the sequences.
}
\author{
Francois Michonneau
}

