% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nexusPartitionCreate}
\alias{nexusPartitionCreate}
\title{Create a NEXUS-formatted partition file}
\usage{
nexusPartitionCreate(alg, file.out = "nexus_partition", part.names = c("p1",
  "p2", "p3"), overwrite = FALSE)
}
\arguments{
\item{alg}{The path to a PHYLIP formatted file}

\item{file.out}{The file name and path to the output file}

\item{part.names}{The names of the partitions used, the length of
this arguments determines how many partitions are generated. By
default 3, making one partition per codon position.}

\item{overwrite}{If a file with the same name as \code{file.out}
exists, should it be overwritten? (default \code{FALSE}.}
}
\value{
a named logical vector indicating whether the file was
created.
}
\description{
Given a PHYLIP file, this function returns a NEXUS formatted
partition file that is split by codon position.
}
\details{
By default the number of partitions is determined by how many
names are provided.

The file used (supplied with \code{alg}) must be a PHYLIP
formatted file: the first line must contain the number of
sequences, followed by the length of the sequences.
}
\author{
Francois Michonneau
}

