% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeEmptySeqs}
\alias{removeEmptySeqs}
\title{Remove empty sequences}
\usage{
removeEmptySeqs(file, output, formatin = "fasta", formatout = formatin,
  colsep = "", overwrite = FALSE, gap, ...)
}
\arguments{
\item{file}{path and file name for a DNA alignment}

\item{output}{path and file name of the alignment with the missing
data removed. If missing, the original file is overwritten, with
an a posteriori warning, unless quiet=TRUE.}

\item{formatin}{format of the alignment to check (to be passed to
\code{\link[ape]{read.dna}})}

\item{formatout}{format of the alignemnt to write (to be passed to
\code{\link[ape]{write.dna}})}

\item{colsep}{argument to be passed to
\code{\link[ape]{write.dna}}, character used to separate columns
in DNA alignment (\code{ape}'s default is a single space, here we
override this to no character, i.e. \code{colsep=""}).}

\item{overwrite}{If no output file is specified and if overwrite is TRUE,
the original file is overwritten.}

\item{gap}{character in the sequence to be considered as missing data}

\item{...}{additional arguments to be passed to
\code{\link[ape]{write.dna}}.}
}
\value{
TRUE if function succeedeed, but mostly used for its side
effect of creating a new alignment file.
}
\description{
Removes empty sequences from an alignment
}
\details{
This function takes an alignment, and remove sequences that only
contains gaps. This function can read the same type of alignment
files as ape:::read.dna, and output the same formats as
ape:::write.dna
}
\author{
Francois Michonneau
}

