% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFastaComments.R
\name{removeFastaComments}
\alias{removeFastaComments}
\title{Remove comments from FASTA files}
\usage{
removeFastaComments(file, output = NULL, ...)
}
\arguments{
\item{file}{The file name of the alignment}

\item{output}{If \code{NULL} (default), the orignal file is
overwritten, otherwise a file name for the output}

\item{...}{Additional arguments to be passed to
\code{\link[ape]{read.dna}}.}
}
\value{
The name of the output file
}
\description{
Remove the comments in the titles of FASTA files (everything that
is after \sQuote{/}).
}
\author{
Francois Michonneau
}

