\name{splitMultiAlignments}
\alias{splitMultiAlignments}
\title{splitMultiAlignments -- split multiple alignments into
individual files}
\usage{
splitMultiAlignments(file, prefix, pathin, pathout)
}
\arguments{
  \item{file}{the name of the file containing the multiple
  alignemnts to be split.}

  \item{prefix}{the prefix for each of the file name}

  \item{pathin}{path (i.e., directory) where the replicates
  for each locus are stored}

  \item{pathout}{path (i.e., directory) where the
  individual replicate for each locus are going to be
  stored.}
}
\value{
TRUE, but really used for its side effect of generating
individual alignment files.
}
\description{
Split a file containing multiple alignments into a multiple
files.
}
\details{
This function reads a file containing multiple alignments
in the phylip format (as generated by seq-gen) and breaks
each replicate into its own file. The files are named by
putting together a prefix (specified by the user), the
string \sQuote{rep} (stands for replicate), the replicate
number, the name of the marker/locus (which should be part
of the file name between an underscore and \sQuote{.phy},
e.g., \code{myalignment_marker.phy}).
}
\author{
Francois Michonneau
}

