% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutAlignment.R
\name{cutAlignment}
\alias{cutAlignment}
\title{Cut an alignment into individual partitions.}
\usage{
cutAlignment(algfile, partfile, formatin = "fasta", formatout = formatin,
  colsep = "", ...)
}
\arguments{
\item{algfile}{file name of the alignment to cut}

\item{partfile}{file name of the partition file}

\item{formatin}{format of the DNA alignment (to be passed to
ape:::read.dna)}

\item{formatout}{format of the DNA alignment to be returned (by
default same as \code{formatin}, can be any value accepted by
\code{\link[ape]{write.dna}})}

\item{colsep}{argument to be passed to
\code{\link[ape]{write.dna}}, character used to separate columns
in DNA alignment (\code{ape}'s default is a single space, here we
override this to no character, i.e. \code{colsep=""}).}

\item{...}{additional arguments controlling the formating of the
output (to be passed to ape:::write.dna)}
}
\value{
TRUE if the function worked, used for its side effect of
creating individual partition files.
}
\description{
Given an alignment and a partition file, takes the alignment and
cuts it into individual partitions.
}
\details{
At this time, only takes a RAxML formatted partition file. The
individual partition generated are named with the full alignment
followed by the partition names specified in the partition file
separated by an underscore. For each of the file created, the same
extension as the full alignment is used. For instance, if your
alignment file is named \code{fullAlignment.phy}, and your
partitions are named \code{part1}, \code{part2} and \code{part3},
this function will generate the files
\code{fullAlignment_part1.phy}, \code{fullAlignment_part2.phy},
and \code{fullAlignment_part3.phy}.
}
\examples{
\dontrun{
cutAlignment("fullAlignment.phy", partfile="partInfo.part",
             formatin="sequential", colsep="")
}
}
\author{
Francois Michonneau
}
\seealso{
\code{\link{write.dna}}
}

