% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitAlignment.R
\name{splitAlignment}
\alias{splitAlignment}
\title{Split alignments.}
\usage{
splitAlignment(file, output, split_by = NULL, format = "fasta",
  colw = 10000, checkAmbiguity = TRUE, checkInternalGaps = FALSE, ...)
}
\arguments{
\item{file}{path and file name of the alignment (all the seq need
to be the same length)}

\item{output}{folder where the individual sequences will be stored}

\item{split_by}{If \code{NULL} (default), each sequence will be a
in a separate file. Otherwise, a data frame with at least 2
columns named \sQuote{groups} and \sQuote{sequences}:
sequences identified in the \sQuote{sequences} column, sharing
the same values in the \sQuote{groups} column will be written
to files.}

\item{format}{format of the alignment, for now only fasta}

\item{colw}{width of columns, same as in
\code{\link[ape]{write.dna}}}

\item{checkAmbiguity}{should the presence of ambiguities in the
sequences be tested? (default=TRUE). If ambiguities are found,
returns a warning.}

\item{checkInternalGaps}{should the presence of internal gaps be
tested? (default=FALSE). If internal gaps are found, returns a
warning.}

\item{...}{additional parameters to be passed to ape:::write.dna}
}
\value{
TRUE if the function succeed, mostly used for its side
    effect (create files with individual sequences).
}
\description{
Split alignments into individual files.
}
\details{
This function takes an alignment and put each sequence into its
individual file that is named the same as the name of the
sequence. This function also overwrites the output (it doesn't
check that the file created don't already exist).
}
\author{
Francois Michonneau
}

